pyproject.toml
[project]
name = "cbir-system"
version = "1.0.0"
description = "Content-Based Image Retrieval System"
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "scikit-image>=0.19.0",
    "matplotlib>=3.5.0",
    "scipy>=1.7.0",
    "pillow>=9.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]